% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_watcher.R
\name{series_update}
\alias{series_update}
\title{Updates the current Series spreadsheet with the results / links of recently
completed games.}
\usage{
series_update(season_key, season_start, pairing_ranges, api_sheet, token)
}
\arguments{
\item{season_key}{The ID ("key") of the current Series season spreadsheet}

\item{season_start}{When the current season started, in YYYY-MM-DD}

\item{pairing_ranges}{The cell references of the ranges in the "API" sheet that have the current season pairings.}

\item{api_sheet}{The name of the worksheet that needs to be updated. Usually "API".}

\item{token}{Your Lichess API token. Must have the msg:write scope.}
}
\value{
Console messages informing you of progress. Final message will read "All done!".
}
\description{
Updates the current Series spreadsheet with the results / links of recently
completed games.
}
\examples{
series_update()
}
